refactor: move `url` and `file_name` functions from source to mod
