+++ src/project/mod.rs
@@ -1,8 +1,10 @@
 pub mod source;
 
 use crate::project::source::Source;
-use eyre::Result;
+use crate::project::source::Source::{Download, Modrinth};
+use eyre::{eyre, Result, WrapErr};
 use std::path::PathBuf;
+use url::Url;
 
 #[derive(Debug, Clone)]
 pub struct Project {
@@ -41,10 +43,52 @@ impl Mod {
         Mod { name, file, source }
     }
 
-    pub fn file_or_source(&self) -> Result<String> {
+    pub fn download_url(&self) -> Result<String> {
+        match &self.source {
+            Download { url } => Ok(url.clone()),
+            Modrinth { version } => Ok(
+                match crate::modrinth::get_version(version)
+                    .wrap_err("failed to fetch modrinth version")?
+                {
+                    Some(version) => version.id,
+                    None => {
+                        crate::modrinth::get_versions(&self.name)
+                            .wrap_err("failed to fetch modrinth project versions")?
+                            .into_iter()
+                            .find(|modrinth_version| &modrinth_version.version_number == version)
+                            .ok_or_else(|| eyre!("could not find version `{}`", version))?
+                            .id
+                    }
+                },
+            ),
+        }
+    }
+
+    pub fn file_name(&self) -> Result<String> {
         match self.file.clone() {
             Some(file) => Ok(file),
-            None => self.source.file_name(&self.name),
+            None => match &self.source {
+                Download { url } => Url::parse(url)?
+                    .path_segments()
+                    .and_then(|segments| segments.last())
+                    .and_then(|name| if name.is_empty() { None } else { Some(name) })
+                    .map(|s| s.into())
+                    .ok_or_else(|| eyre!("invalid url")),
+                Modrinth { version } => Ok(match crate::modrinth::get_version(version)
+                    .wrap_err("failed to fetch modrinth version")?
+                {
+                    Some(version) => version,
+                    None => crate::modrinth::get_versions(&self.name)
+                        .wrap_err("failed to fetch modrinth project versions")?
+                        .into_iter()
+                        .find(|modrinth_version| &modrinth_version.version_number == version)
+                        .ok_or_else(|| eyre!("could not find version `{}`", version))?,
+                }
+                .primary_file()
+                .ok_or_else(|| eyre!("primary file not found"))?
+                .filename
+                .clone()),
+            },
         }
     }
 }
